+++ src/pages/HomePage/HomePage.module.scss
@@ -1,4 +1,4 @@
-@import '../components/global.scss';
+@import '../../components/global.scss';
 
 .mainCardContainer {
   width: 100vw;
