+++ src/pages/Layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@import '../components/global.scss';
+@import '../../components/global.scss';
 
 .mainContainer {
   display: flex;
