+++ src/pages/ViewNote/ViewNote.module.scss
@@ -1,4 +1,4 @@
-@import '../components/global.scss';
+@import '../../components/global.scss';
 
 .note {
   display: grid;
